Reshape UMD to respect module/export first.
Ran into a problem with the UMD wrapper. Here's the scenario:

We wrap libraries in UMD to make them easier to deploy in any environment, however one common environment is CJS to AMD-compiler.

If a library is exported as UMD, but treated as CJS in a local environment, and then compiled to AMD (via a wrapper), then our UMD wrapper misidentifies the environment as AMD, when we want it to behave as CJS. define() is called from within define(), and the outer-most define() ends up thinking its module exports nothing, causing sad trombone music to play out of it's tiny 8-bit speakers.

In other words: AMD defines both "module"/"exports" and "define" where CJS defines only "module"/"exports". However, AMD environments do not globally define "module"/"exports" - only locally to a module definition, in which case they should be treated as CJS treats them.

What this means is that we actually want to detect "module"/"exports" first, falling back to detecting "define" (assuming a global level).